#2 고유값 생성 전략 수정 및 null체크

+++ src/main/java/com/suhyunkim/ticket/follow/FollowBo.java
@@ -27,6 +27,8 @@ public class FollowBo {
 
 	public void unfollow(String userId) {
 		Follower follower = followDao.findByUserId(userId);
-		followDao.delete(follower);
+		if (follower != null) {
+			followDao.delete(follower);
+		}
 	}
 }
